Move client to own module
